==> analysis/02_PreProcessing.Rmd <==
---
title: "Phyloseq PreProcessing"
output: html_document
date: "`r Sys.Date()`"
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align = "center", fig.path = "../figures/02_PreProcessing/") # Send any figures to this folder
```

# Load libraries
```{r load-libraries}
#devtools::install_github("joey711/phyloseq")
pacman::p_load(devtools, phyloseq, tidyverse, dada2, install = FALSE)
```

# Goals

Here, we will process the data into a phyloseq object.

- ASV table
- Taxonomy table
- Track reads (metadata)

Then, we will remove the following:

1. Remove Chloroplasts
2. Remove Mitochondria
3. Remove ASVs from negative controls and also negative control samples.
4. Evaluate accuracy by looking at the Mock Community
5. Remove samples without enough reads.

Finally, we will write a data file of phyloseq output, which will be stored in `data/02_PreProcessing/raw_preprocessed_phyloseq.RData`.

# Load data

## ASV Table
```{r load-ASV-table}
# First, load ASV table
#setwd("/workdir/zlr6/gitrepos/SalinityGradient_16s")
load("/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/01_DADA2/ASV_counts.RData")

# Inspect ASV table
head(asv_tab)[,1:5]

# Fix names
colnames(asv_tab)
sample_names <- colnames(asv_tab)
samples_fixed <- sapply(strsplit(basename(sample_names), "_"), `[`,1)
head(samples_fixed)

# re-write the ASV count file to fix names
colnames(asv_tab) <- samples_fixed
```

## Taxonomy Table
```{r load-tax-table}
tax_df <- read.table("/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/01_DADA2/ASV_taxonomy.tsv", sep = "\t", skip = 1, header = FALSE)

# fix column names
colnames(tax_df) <- c("asv_names", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "ASV", "ASVseq")

# Taxonomy Table Matrix
tax_mat <- tax_df %>%
  tibble::column_to_rownames(., var = "asv_names") %>%
  as.matrix()
```

## Track Reads Data
```{r load-track-reads}
load("/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/01_DADA2/track_read_counts.RData")
# Take a look at the data
head(track_counts_df)
dim(track_counts_df)

# Load in metadata
metadata_df <- read.csv("/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/metadata.csv")
dim(metadata_df)
colnames(metadata_df)

# Merge metadata_df with track_reads_df
head(metadata_df)
head(track_counts_df)

metadata_track_reads_df <- metadata_df %>%
  left_join(., track_counts_df, by = "names") %>% dplyr::select(-X)

# Intuition check
head(metadata_track_reads_df)

# Update row.names to be sample names
# Before
row.names(metadata_track_reads_df)

# Rewrite
row.names(metadata_track_reads_df) <- metadata_track_reads_df$names

# Check afterwards that it worked
row.names(metadata_track_reads_df)

# Intuition check
head(metadata_track_reads_df)
```

# Handoff
```{r phyloseq-handoff}
# Double-check it's all good
dim(asv_tab)
dim(tax_mat)

# Intuition check
stopifnot(row.names(asv_tab) == row.names(tax_mat))

# Construct the phyloseq object
raw_physeq <- phyloseq(otu_table(asv_tab, taxa_are_rows = TRUE), sample_data(metadata_track_reads_df), tax_table(tax_mat))

raw_physeq

# Save this raw phyloseq object
save(raw_physeq, file = "data/02_PreProcessing/raw_physeq.RData")
```

# Clean up the data

Remove:

1. Chloroplasts
2. Mitochondria

```{r rm-mitos-chloros}

# Make new physeq without mitochondria or chloroplasts
noChloros_physeq <- raw_physeq %>% 
  # rm chloroplasts
  subset_taxa(Order != "Chloroplast" | is.na(Order))

# How many taxa were chloroplasts?
num_chloro_ASVs <- ntaxa(raw_physeq) - ntaxa(noChloros_physeq)
num_chloro_ASVs

# Intuition check
# noChloros_physeq %>%
#  tax_table() %>%
#  data.frame() %>%
#  View()

# Remove mitochondria
# Make new physeq without mitochondria or chloroplasts
noChlorosMitos_physeq <- noChloros_physeq %>% 
  # rm mitochondria
  subset_taxa(Family != "Mitochondria" | is.na(Family))
noChlorosMitos_physeq

# How many taxa were mitochondria?
num_mitos_ASVs <- ntaxa(noChloros_physeq) - ntaxa(noChlorosMitos_physeq)
num_chloro_ASVs
num_mitos_ASVs

# How many total ASVs were removed from Chloros and Mitos
ntaxa(raw_physeq) - ntaxa(noChlorosMitos_physeq)
# proportion of asvs kept?
ntaxa(noChlorosMitos_physeq)/ntaxa(raw_physeq)
```

# Evaluate and remove the control samples

Take a look at the negative controls and then make a decision about whether or not to remove the ASVs that we found in our controls

1. Negative controls
2. ASVs found within the negative contorls and their distribution in the samples.
3. Evaluate the mock community

# Negative Controls

```{r neg-controls}
# create a vector of samples that were negative controls
control_samples <- c("WaterControl", "022um-Control", "3um-Control", "DNA-Ext-Control")

# Make a new phyloseq object of only the control samples
control_physeq <- noChlorosMitos_physeq %>% 
  subset_samples(., names %in% control_samples) %>%
  # force remove ASVs with 0 counts
  prune_taxa(taxa_sums(.) > 0, .)

# vector of control asvs
control_ASVs <-
  control_physeq %>%
  tax_table() %>%
  data.frame() %>%
  dplyr::select(ASV) %>%
  as.vector()

# Visualize raw abundance values of control ASVs
control_physeq %>%
  plot_bar(., "ASV", fill = "ASV")

# Check control AVSs in the. samples
# Make new phyloseq object with only the control ASVs and ALL samples
control_ASV_physeq <- 
  noChlorosMitos_physeq %>%
  subset_taxa(., ASV %in% control_ASVs$ASV)
control_ASV_physeq

# Let's take a look at the abundances
control_ASV_physeq %>%
  plot_bar(., "ASV", fill = "ASV", facet_grid = Sample_or_Control~.)

# Check tax of ASV_3
control_ASV_physeq %>%
  tax_table() %>%
  data.frame()

# Would only like to keep ASV_18 and ASV_56, everything else is trash
controlASVs_toREMOVE <- dplyr::setdiff(control_ASVs$ASV, c("ASV_18", "ASV_56"))
# double check
length(controlASVs_toREMOVE)

# make my new phyloseq object
noControlsChlorosMitos_physeq <- 
  noChlorosMitos_physeq %>%
  # Actually remove control ASVs (except 18, 56)
  subset_taxa(., !(ASV %in% controlASVs_toREMOVE)) %>%
  # Remove control samples
  subset_samples(., !(names %in% control_samples))

# Inspect no control object
noControlsChlorosMitos_physeq

# After filtering, how many ASVs were removed?
total_ASVs_rm <- ntaxa(raw_physeq) - ntaxa(noControlsChlorosMitos_physeq)
total_ASVs_rm
```

# Evaluate accuracy in the Mock Community

- make a mock phyloseq object
- load in the standard fasta sequences of the mock community
- test for differences between the two

```{r mock-phyloseq}
# Make mock phyloseq object
mock_physeq <- noControlsChlorosMitos_physeq %>%
  subset_samples(., names == "MockZymoPos") %>%
  prune_taxa(taxa_sums(.) > 0, .)

# Insoect it
tax_table(mock_physeq)

# write us a message summarizing this
cat("DADA2 inferred", ntaxa(mock_physeq), "ASVs present in mock community.")

##### Load in the mock reference sequence from Zymo
mock_ref <- getSequences("/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/02_PreProcessing/mock_amplicons.fasta")
names(mock_ref)

## Pull out seqs from Sequenced Mock community
mock_seqs <- 
  mock_physeq %>%
  tax_table() %>%
  data.frame() %>%
  dplyr::select(ASV, ASVseq)
# Inspect
head(mock_seqs)

# Which ASVs match the reference?
matched_ASVs <- 
  mock_seqs %>%
  rowwise() %>%
  # Check each ASV, if they match, TRUE, if not = FALSE
  mutate(Match = any(grepl(ASVseq, mock_ref))) %>%
  # Create a vector of names that matched
  pull(ASV)

cat(length(matched_ASVs), "ASVs were exact matches to the expeted reference sequence.")

# How many sequences were in the mock?
mock_SeqDepth <- 
  mock_physeq %>%
  otu_table() %>%
  data.frame() %>%
  colSums()

# Who is in the sequenced mock community?
mock_physeq %>%
  tax_table() %>%
  data.frame() %>%
  dplyr::select(Genus, Species)
names(mock_ref)

# Curious: What are the mock counts in the actual samples

```

# Remove mock ASVs
```{r rm-mock-asvs}
# Make a new phyloseq object without mock and its asvs
noMockControlsChlorosMitos_physeq <-  noControlsChlorosMitos_physeq %>%
  # remove mock community
  subset_samples(., names != "MockZymoPos") %>%
  # remove ASVs from the mock
  subset_taxa(., !(ASV %in% matched_ASVs))
```

# Evaluate the sequencing depth

```{r}
# The current data object
noControlsChlorosMitos_physeq

# What is the library size/sequencing depth for each sample?
seqSums_df <- 
  noMockControlsChlorosMitos_physeq %>%
  otu_table() %>%
  # sum each sample column
  colSums() %>%
  data.frame() %>%
  rownames_to_column(var = "names") %>%
  left_join(., metadata_track_reads_df, by = "names")

# check
dim(seqSums_df)

# Rename second column
colnames(seqSums_df)[2] <- "TotalSeqs"

# plot it!
seqSums_df %>%
  ggplot(aes(x=reorder(names, TotalSeqs), y=TotalSeqs, fill = station)) +
  geom_bar(stat = "identity")

# Density plot
seqSums_df %>%
  ggplot(aes(TotalSeqs)) +
  geom_density()

# Show the depth of sampels
seqSums_df %>%
  dply
```

```{r rm-samps}
# what's the min seq depth?
min(sample_sums(noMockControlsChlorosMitos_physeq))

# remove 20210615
raw_preprocessed_physeq <-
  noMockControlsChlorosMitos_physeq %>%
  subset_samples(., names != "20210615-MA-SCB2F")

# what's the new min seq depth?
min(sample_sums(raw_preprocessed_physeq))
```

```{r save-physeq}
save(raw_preprocessed_physeq,
     file = "/local/workdir/zlr6/git_repos/SalinityGradient_16s/data/02_PreProcessing/raw_preprocessed_physeq.RData")
```

# Session Information 
```{r session-info}
# Ensure reproducibility 
devtools::session_info()
```
