==> analysis/04_Biodiversity.Rmd <==
---
title: "04_Biodiversity"
output: 
  html_document:
    toc: yes
    toc_float:
      collapse: no
      smooth_scroll: yes
      toc_depth: 3
    theme: "default"
date: "`r Sys.Date()`"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align = "center", fig.path = "../figures/04_Biodiversity/") # Send any figures to this folder
```

# Setting the Environment

## Set the seed
```{r set-seed}
set.seed(31415)
```

## Load Libraries
```{r load-packages}
#install.packages("iNEXT")
pacman::p_load(devtools, phyloseq, patchwork, iNEXT, tidyverse, install = FALSE)
```

## Load in Data
```{r load-data}
load("/workdir/zlr6/git_repos/SalinityGradient_16s/data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intuition check
min(sample_sums(raw_preprocessed_physeq))
```

# Goals

1. Calculate the Hill Diversity of the samples.
2. Evaluate the rarefaction curves.
3. Evaluate the diversity values.
4. Make notes of specific samples and their seq depth.

# Diversity Calculations wirth iNEXT
```{r calc-div}
# prepare input data
iNEXT_input_df <- raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

# Quick check
dim(iNEXT_input_df)

# Run iNEXT: Calculate the Hill numbers
# Note that species in ROWS, samples in COLUMNS
# Remember to set the seed!
#iNEXT_data <- iNEXT(iNEXT_input_df, 
                   # q = c(0, 1, 2), datatype = "abundance")

# save the file
#save(iNEXT_data, file = "/workdir/zlr6/git_repos/SalinityGradient_16s/data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the diversity!
```{r load-div}
load("/workdir/zlr6/git_repos/SalinityGradient_16s/data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
metadata_df <- read.csv("/workdir/zlr6/git_repos/SalinityGradient_16s/data/metadata.csv")
```

```{r plot-rarefaction}
# setting colors for stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")

# Prepare colors
color_df <-
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
#check
head(color_df)
# rename the column
colnames(color_df)[1] <- "names"
#check
head(color_df)

# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
```

# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) + 
  theme(legend.position = "none")
```

# Manually plot diversity

## Rarefaction

```{r iNEXT-manual}
iNEXT_manual_df <- iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # fix the sample names
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  # join with metadata
  left_join(., metadata_df, by = "names") %>%
  # add colors to data frame
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)),
            by = "station")

# plot it! - rarefaction curve
iNEXT_manual_df %>%
  # filter out rows that are calculated by rarefaction from iNEXT
  dplyr::filter(Method == "Extrapolation") %>%
  # make the actual plot
  ggplot(aes(x = m, y = qD, color = station, group = names)) +
  geom_line(alpha = 0.5) +
  facet_grid(Order.q~station, scales = "fixed") +
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")
```


# Environmental variable check

```{r environmental-pairs, fig.width=8, fig.height=8}
head(metadata_df)
# Pull out environmental variables
env_df <-
    metadata_df %>%
    dplyr::select(names, water_tempC:DO_mgL)
# inspect
head(env_df)

# plot the correlation
pairs(dplyr::select(env_df, -names), upper.panel = NULL)
```

# Diveristy vs salinity

```{r div-vs-salinity, fig.height=6, fig.width=6}
iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = salinity_psu, y = qD)) +
  facet_wrap(.~Order.q, scales = "free") +
  stat_smooth(method = "lm", formula = y ~poly(x, 2)) +
  geom_point(aes(color = station)) +
  scale_color_manual(values = station_colors) +
  labs(x = "Salinity (PSU)", y = "# of ASVs") +
  theme(legend.position = "bottom")
```

------------------------------------------------------

# Session Information
```{r session-info}
devtools::session_info()
```
